requirements.txt
ultralytics==8.3.229
Pillow==12.0.0
PyYAML==6.0.3
